requirements.txt
# Web API framework
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OCR tools
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=9.5.0

# Utils
python-dotenv>=1.0.0